requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
setuptools==60.5.0
cffi==1.14.5
DAWG-Python==0.7.2
Django==3.0.6
django-cors-headers==3.2.0
django-crontab==0.7.1
djangorestframework==3.11.1
docopt==0.6.2
drf-yasg==1.20.0
exponent-server-sdk==2.0.0
funcy==1.15
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==9.3.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.4.7
pyTelegramBotAPI==3.8.2
pytz==2021.1
pyjwt==2.3.0
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
boto3==1.22.5
meilisearch==0.18.2
django-jsonfield==1.4.1

